fix: use f64 for coords instead of f32.

+++ src/coord.rs
@@ -3,8 +3,8 @@ use geo::algorithm::contains::Contains;
 use geo::{Point, Polygon};
 use std::collections::HashMap;
 pub struct Coord {
-    lat: f32,
-    lng: f32,
+    lat: f64,
+    lng: f64,
 }
 
 impl Coord {
@@ -12,8 +12,8 @@ impl Coord {
         let items: Vec<&str> = input.split(",").collect();
         match items.len() {
             2 => Ok(Coord {
-                lat: items[0].parse::<f32>()?,
-                lng: items[1].parse::<f32>()?,
+                lat: items[0].parse::<f64>()?,
+                lng: items[1].parse::<f64>()?,
             }),
             _ => bail!("need 2 float for coordinate"),
         }
@@ -30,19 +30,19 @@ impl Coord {
 }
 
 impl Locatable for Coord {
-    fn lat(&self) -> f32 {
+    fn lat(&self) -> f64 {
         self.lat
     }
-    fn lng(&self) -> f32 {
+    fn lng(&self) -> f64 {
         self.lng
     }
 }
 
 pub trait Locatable {
-    fn lat(&self) -> f32;
-    fn lng(&self) -> f32;
-    fn locate<'a>(&self, areas: &'a HashMap<String, Vec<Polygon<f32>>>) -> Result<&'a String> {
-        let p = Point::<f32>::new(self.lng(), self.lat());
+    fn lat(&self) -> f64;
+    fn lng(&self) -> f64;
+    fn locate<'a>(&self, areas: &'a HashMap<String, Vec<Polygon<f64>>>) -> Result<&'a String> {
+        let p = Point::<f64>::new(self.lng(), self.lat());
         for (k, vs) in areas.iter() {
             for v in vs {
                 if v.contains(&p) {
